+++ src/main/java/Seed/Model/BlockMeta.java
@@ -4,10 +4,10 @@ import java.io.Serializable;
 
 public class BlockMeta implements Serializable {
     final int id;
-    public final int size;
+    public final long size;
     public final byte[] data;
 
-    public BlockMeta(int id, int size, byte[] data) {
+    public BlockMeta(int id, long size, byte[] data) {
         this.id = id;
         this.size = size;
         this.data = data;
